chore: move db file to utils
