+++ src/routes/auth/data.ts
@@ -1,5 +1,5 @@
-import { DBTables } from '../../constants/db.ts';
 import type { AccessLevel } from '../../utils/auth.ts';
+import { DBTables } from '../../utils/db.ts';
 
 export async function getLoginInfo(database: D1Database, email: string) {
 	const loginInfo = await database
