+++ src/routes/profile/validation.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { BaseDbEntitySchema, BaseDBFieldsToOmit } from '../../constants/db.ts';
+import { BaseDbEntitySchema, BaseDBFieldsToOmit } from '../../utils/db.ts';
 
 export const ProfileSchema = BaseDbEntitySchema.merge(z.object({
 	email: z
