+++ src/routes/team/index.ts
@@ -1,8 +1,8 @@
 import { createRoute, OpenAPIHono } from '@hono/zod-openapi';
 import { authMiddleware } from 'src/middleware/auth.ts';
 import { z } from 'zod';
-import { DBTables } from '../../constants/db.ts';
 import { authorizeAdmin, authorizeOrganizer } from '../../middleware/access.ts';
+import { DBTables } from '../../utils/db.ts';
 import {
 	type DataResponse,
 	generateDataResponeSchema,
