+++ src/utils/db.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import type { AccessLevel } from '../utils/auth.ts';
+import type { AccessLevel } from './auth.ts';
 
 export const SCHEMA_VERSION = 1;
 export const DEFAULT_TEAM_ID = 'b3410598-ecbc-41be-9f68-925da74bc613';
